strutil.c: Translate ` ` to `\032` in .map file symbol values
Sometime in 2008 the documented behaviour of outputting all spaces in
symbol values in map files as `\032` was lost. I contacted the author
about this and he sent me a patch, which is this commit.

This commit can be dropped when a new version with the patch is
released, which will presumably be asl-current-142-bld151.

Patch from Alfred Arnold <alfred@ccac.rwth-aachen.de> via email.