download.sh: Fix check for unclean repo
The previous check did not actually exit if there were uncommited files.
(I'm not clear on why, but the current code is rather more clear anyway.)

There are various other fixes that also need to go in, starting with
changing `set -e` to `set -eu -o pipefail` and fixing the resulting errors.