Makefile.def: Remove "1.4" from version numbers in version.c
This affects old versions of ASL, from before the *DebVersion string
in version.c was converted to AS_VERSION_* values in version.h.

For the new version.h code, we return only `bld###` because why bother to
parse the other two variables to get the "1.42" that's always the same
anyway? But we weren't removing the "1.42" from the string we got from
version.c, causing older and newer versions to get differently formatted
install dirs. This fixes the older to match the newer, since that's the
easier fix.