Makefile.def: Remove -march=athlon64 so this builds on ARM
Though ARM has an -march flag, it of course does not support setting that
to athlon64. (We had this here probably because it was copied from the
x86_64-specific Makefile.def in the samples.) By default, we probably don't
want to bother using -march at all, since it's not likely to give any huge
(or noticable) speed improvement given how fast this already runs on modern
machines, but we do leave in a commented-out -march=native (which should
always be accepted on all CPUs, even if doesn't do anything) in case
someone wants to try it out.

For further options info, see:
• https://gcc.gnu.org/onlinedocs/gcc/x86-Options.html
• https://gcc.gnu.org/onlinedocs/gcc/ARM-Options.html
• https://gcc.gnu.org/onlinedocs/gcc/Submodel-Options.html