README,gitignore: Add gitignore for upstream branch; improve instructions
We now first suggest using the Makefile.def from the `main` branch, giving
a `git show` command that will retrive it without having to check out that
branch. We also now supply a `gitignore` file that makes `git status` in
the upstream branch much cleaner.

We continue to mention the upstream-supplied `Makefile.def` files in
the samples directory.