BROKEN .github/issue_template: Note about our code vs ASL code and mailing list
Unfortunately this doesn't work. I've tried different names, putting it
under .github/ISSUE_TEMPLATES/, etc. but it still doesn't come up when
someone starts a new issue.

GitHub has some amazingly complex web interface stuff for trying to
set this up, but the best I've gotten from that was to end up with
the option to use a blank template or the issue template (selecting
from a list with those two options), which is not what we want.

What we actually want is just this template appearing in the description
of any PR or issue that someone starts creating.